unify cargo version and add some ci rules
