+++ clippy.toml
@@ -1,0 +1,11 @@
+disallowed-types = [
+    { path = "once_cell::sync::Lazy", reason = "Please use `std::sync::LazyLock` instead." },
+]
+
+disallowed-macros = [
+    { path = "lazy_static::lazy_static", reason = "Please use `std::sync::LazyLock` instead." },
+]
+
+too-many-arguments-threshold = 10
+upper-case-acronyms-aggressive = false
+enum-variant-size-threshold = 200
