+++ rust-toolchain.toml
@@ -1,0 +1,3 @@
+[toolchain]
+channel = "nightly-2025-10-01"
+components = ["rustfmt", "clippy", "rust-src", "miri", "rust-analyzer"]
