+++ rustfmt.toml
@@ -1,0 +1,9 @@
+edition = "2024"
+style_edition = "2024"
+reorder_imports = true
+group_imports = "StdExternalCrate"
+where_single_line = true
+trailing_comma = "Vertical"
+overflow_delimited_expr = true
+format_code_in_doc_comments = true
+normalize_comments = true
