+++ src/lib.rs
@@ -1,6 +1,3 @@
-use crate::models::{minicpm4::generate::MiniCPMGenerateModel, qwen2_5vl::generate::Qwen2_5VLGenerateModel, GenerateModel};
-use anyhow::{Ok, Result};
-use candle_core::{DType, Device};
 pub mod chat_template;
 pub mod models;
 pub mod position_embed;
