+++ src/models/minicpm4/generate.rs
@@ -1,13 +1,3 @@
-use crate::models::minicpm4::config::MiniCPM4Config;
-use crate::models::minicpm4::model::MiniCPMModel;
-// use crate::models::GenerateStream;
-use crate::utils::utils::{
-    build_completion_chunk_response, build_completion_response, find_type_files, get_device, get_dtype, get_logit_processor
-};
-use crate::{
-    chat_template::chat_template::ChatTemplate, models::GenerateModel,
-    tokenizer::tokenizer::TokenizerModel,
-};
 use anyhow::{Result, anyhow};
 use candle_core::{DType, Device, Tensor};
 use candle_nn::VarBuilder;
@@ -16,6 +6,15 @@ use openai_dive::v1::resources::chat::{
 };
 use rocket::async_stream::stream;
 use rocket::futures::Stream;
+
+use crate::models::minicpm4::config::MiniCPM4Config;
+use crate::models::minicpm4::model::MiniCPMModel;
+// use crate::models::GenerateStream;
+use crate::utils::{
+    build_completion_chunk_response, build_completion_response, find_type_files, get_device,
+    get_dtype, get_logit_processor,
+};
+use crate::{chat_template::ChatTemplate, models::GenerateModel, tokenizer::TokenizerModel};
 
 pub struct MiniCPMGenerateModel<'a> {
     chat_template: ChatTemplate<'a>,
@@ -26,7 +25,7 @@ pub struct MiniCPMGenerateModel<'a> {
     im_end_id: u32,
 }
 
-impl <'a> MiniCPMGenerateModel<'a> {
+impl<'a> MiniCPMGenerateModel<'a> {
     pub fn init(path: &str, device: Option<&Device>, dtype: Option<DType>) -> Result<Self> {
         let chat_template = ChatTemplate::init(path)?;
         let tokenizer = TokenizerModel::init(path)?;
@@ -37,7 +36,7 @@ impl <'a> MiniCPMGenerateModel<'a> {
         let dtype = get_dtype(dtype, cfg_dtype);
         let endoftext_id = cfg.eos_token_id[0];
         let im_end_id = cfg.eos_token_id[1];
-        let model_list = find_type_files(&path, "safetensors")?;
+        let model_list = find_type_files(path, "safetensors")?;
         let vb = unsafe { VarBuilder::from_mmaped_safetensors(&model_list, dtype, device)? };
         let minicpm = MiniCPMModel::new(vb, cfg)?;
 
@@ -53,7 +52,6 @@ impl <'a> MiniCPMGenerateModel<'a> {
 }
 
 impl<'a> GenerateModel for MiniCPMGenerateModel<'a> {
-
     fn generate(&mut self, mes: ChatCompletionParameters) -> Result<ChatCompletionResponse> {
         let mut logit_processor = get_logit_processor(mes.temperature, mes.top_p);
         let mes_render = self.chat_template.apply_chat_template(&mes)?;
@@ -61,10 +59,7 @@ impl<'a> GenerateModel for MiniCPMGenerateModel<'a> {
         let mut seq_len = input_ids.dim(1)?;
         let mut seqlen_offset = 0;
         let mut generate = Vec::new();
-        let sample_len = match mes.max_tokens {
-            Some(max) => max,
-            None => 2048,
-        };
+        let sample_len = mes.max_tokens.unwrap_or(2048);
         for _ in 0..sample_len {
             let logits = self.minicpm.forward_with_cache(&input_ids, seqlen_offset)?;
             let logits = logits.squeeze(0)?.squeeze(0)?.to_dtype(DType::F32)?;
@@ -91,10 +86,7 @@ impl<'a> GenerateModel for MiniCPMGenerateModel<'a> {
         let mut input_ids = self.tokenizer.text_encode(mes_render, &self.device)?;
         let mut seq_len = input_ids.dim(1)?;
         let mut seqlen_offset = 0;
-        let sample_len = match mes.max_tokens {
-            Some(max) => max,
-            None => 512,
-        };
+        let sample_len = mes.max_tokens.unwrap_or(512);
         let stream = stream! {
             let mut error_tokens = Vec::new();
             for _ in 0..sample_len {
@@ -105,7 +97,7 @@ impl<'a> GenerateModel for MiniCPMGenerateModel<'a> {
                 let logits = logits.squeeze(0)?.squeeze(0)?.to_dtype(DType::F32)?;
                 let next_token = logit_processor.sample(&logits)?;
                 let mut decode_ids = Vec::new();
-                if error_tokens.len() > 0 {
+                if !error_tokens.is_empty(){
                     decode_ids.extend_from_slice(&error_tokens);
                 }
                 decode_ids.push(next_token);
