+++ src/models/minicpm4/mod.rs
@@ -1,3 +1,3 @@
 pub mod config;
-pub mod model;
 pub mod generate;
+pub mod model;
