+++ src/models/voxcpm/mod.rs
@@ -1,6 +1,6 @@
-pub mod config;
 pub mod audio_vae;
-pub mod minicpm4;
-pub mod tokenizer;
-pub mod model;
+pub mod config;
 pub mod generate;
+pub mod minicpm4;
+pub mod model;
+pub mod tokenizer;
